services/pipeline/pyproject.toml
[project]
name = "estategap-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13",
    "asyncpg>=0.29",
    "cachetools>=5.0",
    "geopandas>=0.14",
    "geoalchemy2>=0.14",
    "httpx>=0.27",
    "lxml>=5.0",
    "nats-py>=2.6",
    "pyosmium>=3.7",
    "prometheus-client>=0.20",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "rapidfuzz>=3.9",
    "shapely>=2.0",
    "sqlalchemy>=2.0",
    "estategap-common",
    "structlog>=24.1",
]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-httpx>=0.30",
    "ruff",
    "testcontainers[postgres]>=4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["estategap_pipeline", "src/pipeline"]

[tool.uv.sources]
estategap-common = { path = "../../libs/common", editable = true }

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = ["src"]
